fortune-teller/requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
Pillow==10.0.0